added openvr to resoure and install rule to pom, also changed to default directory structure
