+++ pom.xml
@@ -1,27 +1,57 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>sc.fiji</groupId>
-  <artifactId>ThreeDViewer</artifactId>
-  <version>0.0.1-SNAPSHOT</version>
-  <build>
-    <sourceDirectory>src</sourceDirectory>
-    <plugins>
-      <plugin>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.3</version>
-        <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-  <parent>
-  	<groupId>net.imagej</groupId>
-  	<artifactId>pom-imagej</artifactId>
-  	<version>15.8.0</version>
-  </parent>
-  <repositories>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<groupId>sc.fiji</groupId>
+	<artifactId>ThreeDViewer</artifactId>
+	<version>0.0.1-SNAPSHOT</version>
+	<build>
+		<plugins>
+			<plugin>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<version>3.3</version>
+				<configuration>
+					<source>1.8</source>
+					<target>1.8</target>
+				</configuration>
+			</plugin>
+
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-install-plugin</artifactId>
+				<version>2.5.2</version>
+				<executions>
+					<execution>
+						<id>install-external</id>
+						<phase>clean</phase>
+						<configuration>
+
+							<file>${basedir}/src/main/resources/openvr-0.1.jar</file>
+							<repositoryLayout>default</repositoryLayout>
+							<groupId>scenery</groupId>
+							<artifactId>openvr</artifactId>
+							<version>0.1</version>
+							<packaging>jar</packaging>
+							<generatePom>true</generatePom>
+						</configuration>
+						<goals>
+							<goal>install-file</goal>
+						</goals>
+					</execution>
+				</executions>
+			</plugin>
+		</plugins>
+	</build>
+	
+	<properties>
+		<enforcer.skip>true</enforcer.skip>
+	</properties>
+	
+	<parent>
+		<groupId>net.imagej</groupId>
+		<artifactId>pom-imagej</artifactId>
+		<version>15.8.0</version>
+	</parent>
+	<repositories>
 		<repository>
 			<id>imagej.public</id>
 			<url>http://maven.imagej.net/content/groups/public</url>
@@ -35,77 +65,77 @@
 			<url>http://nativelibs4java.sourceforge.net/maven</url>
 		</repository>
 	</repositories>
-  <dependencies>
-  	<dependency>
-  		<groupId>net.clearvolume</groupId>
-  		<artifactId>scenery</artifactId>
-  		<version>1.0-SNAPSHOT</version>
-  	</dependency>
-  	<dependency>
-  		<groupId>net.clearvolume</groupId>
-  		<artifactId>cleargl</artifactId>
-  		<version>2.0.0-SNAPSHOT</version>
-  	</dependency>
-  	<dependency>
-  		<groupId>scenery</groupId>
-  		<artifactId>openvr</artifactId>
-  		<version>0.1</version>
-  	</dependency>
-  	<dependency>
-  		<groupId>net.coremem</groupId>
-  		<artifactId>CoreMem</artifactId>
-  		<version>0.1.5</version>
-  	</dependency>
-  	<dependency>
-  		<groupId>org.jetbrains.kotlin</groupId>
-  		<artifactId>kotlin-stdlib</artifactId>
-  		<version>1.0.2</version>
-  	</dependency>
-  	<dependency>
-  		<groupId>org.scijava</groupId>
-  		<artifactId>scijava-common</artifactId>
-  	</dependency>
-  	<dependency>
-  		<groupId>org.scijava</groupId>
-  		<artifactId>scijava-ui-swing</artifactId>
-  	</dependency>
-  	<dependency>
-  		<groupId>net.imagej</groupId>
-  		<artifactId>imagej-common</artifactId>
-  	</dependency>
-  	<dependency>
-  		<groupId>net.imagej</groupId>
-  		<artifactId>imagej-deprecated</artifactId>
-  	</dependency>
-  	<dependency>
-  		<groupId>net.imagej</groupId>
-  		<artifactId>ij</artifactId>
-  	</dependency>
-  	<dependency>
-  		<groupId>net.imagej</groupId>
-  		<artifactId>imagej-ui-swing</artifactId>
-  	</dependency>
-  	<dependency>
-  		<groupId>net.imagej</groupId>
-  		<artifactId>imagej-ops</artifactId>
-  		<version>0.30.1-SNAPSHOT</version>
-  	</dependency>
-  	<dependency>
-  		<groupId>org.scijava</groupId>
-  		<artifactId>script-editor</artifactId>
-  	</dependency>
-  	<dependency>
-    	<groupId>net.java.jinput</groupId>
-	    <artifactId>jinput</artifactId>
-	    <version>2.0.6</version>
-	</dependency>
-  	<dependency>
-  		<groupId>org.scijava</groupId>
-  		<artifactId>native-lib-loader</artifactId>
-  	</dependency>
-  	<dependency>
-  		<groupId>net.imglib2</groupId>
-  		<artifactId>imglib2-ij</artifactId>
-  	</dependency>
-  </dependencies>
+	<dependencies>
+		<dependency>
+			<groupId>net.clearvolume</groupId>
+			<artifactId>scenery</artifactId>
+			<version>1.0-SNAPSHOT</version>
+		</dependency>
+		<dependency>
+			<groupId>net.clearvolume</groupId>
+			<artifactId>cleargl</artifactId>
+			<version>2.0.0-SNAPSHOT</version>
+		</dependency>
+		<dependency>
+			<groupId>scenery</groupId>
+			<artifactId>openvr</artifactId>
+			<version>0.1</version>
+		</dependency>
+		<dependency>
+			<groupId>net.coremem</groupId>
+			<artifactId>CoreMem</artifactId>
+			<version>0.1.5</version>
+		</dependency>
+		<dependency>
+			<groupId>org.jetbrains.kotlin</groupId>
+			<artifactId>kotlin-stdlib</artifactId>
+			<version>1.0.2</version>
+		</dependency>
+		<dependency>
+			<groupId>org.scijava</groupId>
+			<artifactId>scijava-common</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.scijava</groupId>
+			<artifactId>scijava-ui-swing</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>net.imagej</groupId>
+			<artifactId>imagej-common</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>net.imagej</groupId>
+			<artifactId>imagej-deprecated</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>net.imagej</groupId>
+			<artifactId>ij</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>net.imagej</groupId>
+			<artifactId>imagej-ui-swing</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>net.imagej</groupId>
+			<artifactId>imagej-ops</artifactId>
+			<version>0.30.1-SNAPSHOT</version>
+		</dependency>
+		<dependency>
+			<groupId>org.scijava</groupId>
+			<artifactId>script-editor</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>net.java.jinput</groupId>
+			<artifactId>jinput</artifactId>
+			<version>2.0.6</version>
+		</dependency>
+		<dependency>
+			<groupId>org.scijava</groupId>
+			<artifactId>native-lib-loader</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>net.imglib2</groupId>
+			<artifactId>imglib2-ij</artifactId>
+		</dependency>
+	</dependencies>
 </project>
